🐛 #106 :: Null 처리 버그 수정

+++ src/main/java/com/example/mukgen/domain/review/service/ReviewService.java
@@ -201,7 +201,7 @@ public class ReviewService {
 
         String profileUrl = "";
 
-        if(!review.getImageUrl().isEmpty()){
+        if(review.getImageUrl() != null && !review.getImageUrl().isEmpty()){
             profileUrl = review.getImageUrl();
             s3Util.deleteFile(profileUrl.split("/")[3]);
         }
